Added fhir code to link.py; added tests
