+++ phdi/fhir/linkage/__init__.py
@@ -1,0 +1,3 @@
+from .link import add_patient_identifier
+
+__all__ = ["add_patient_identifier"]
