+++ phdi/fhir/linkage/link.py
@@ -1,9 +1,10 @@
-import hashlib
 import copy
 from phdi.fhir.utils import (
     find_resource_by_type,
     get_field,
+    get_one_line_address,
 )
+from ...linkage.link import generate_hash_str
 
 
 def add_patient_identifier(bundle: dict, salt_str: str, overwrite: bool = True) -> dict:
@@ -37,7 +38,7 @@ def add_patient_identifier(bundle: dict, salt_str: str, overwrite: bool = True) 
         address_line = ""
         if "address" in patient:
             address = get_field(patient, "address", "home", 0)
-            address_line = _get_one_line_address(address)
+            address_line = get_one_line_address(address)
 
         # TODO Determine if minimum quality criteria should be included, such as min
         # number of characters in last name, valid birth date, or address line
@@ -63,20 +64,3 @@ def add_patient_identifier(bundle: dict, salt_str: str, overwrite: bool = True) 
         )
 
     return bundle
-
-
-def generate_hash_str(linking_identifier: str, salt_str: str) -> str:
-    """
-    Given a string made of concatenated patient information, generate
-    a hash for this string to serve as a "unique" identifier for the
-    patient.
-
-    :param linking_identifier: The concatenation of a patient's name,
-        address, and date of birth, delimited by dashes
-    :param salt_str: The salt to concatenate onto the end to prevent
-        being able to reverse-engineer PII
-    """
-    hash_obj = hashlib.sha256()
-    to_encode = (linking_identifier + salt_str).encode("utf-8")
-    hash_obj.update(to_encode)
-    return hash_obj.hexdigest()
